backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
supabase>=2.3.0
stable-baselines3>=2.2.1
gymnasium>=0.29.1
transformers>=4.40.0
python-dotenv>=1.0.0
faker>=24.11.0
pandas>=2.2.0
numpy>=1.26.0
torch>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
pydantic>=2.0.0
